requirements.txt
# Core ML and Data Science
torch>=1.12.0
torch-audio>=0.12.0
torch-vision>=0.13.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
scipy>=1.8.0

# NLP and Sentiment Analysis
transformers>=4.20.0
sentencepiece>=0.1.96
tokenizers>=0.12.0
nltk>=3.7
spacy>=3.4.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Data Processing and Features
ta>=0.10.2
yfinance>=0.1.87
alpha-vantage>=2.3.1
quandl>=3.7.0
feedparser>=6.0.10
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Real-time Data and Messaging
kafka-python>=2.0.2
redis>=4.3.0
websockets>=10.3
aiohttp>=3.8.0
asyncio>=3.4.3

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.8.0

# Analytics and Visualization
plotly>=5.10.0
dash>=2.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
tableauserverclient>=0.17.0

# Backtesting and Risk
zipline-reloaded>=2.2.0
empyrical>=0.5.5
pyfolio>=0.9.2

# Utilities
pyyaml>=6.0
python-dotenv>=0.20.0
click>=8.1.0
tqdm>=4.64.0
joblib>=1.1.0

# API and Web
fastapi>=0.79.0
uvicorn>=0.18.0
pydantic>=1.9.0
requests>=2.28.0
aiofiles>=0.8.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-asyncio>=0.19.0
pytest-mock>=3.8.0

# Code Quality
black>=22.6.0
isort>=5.10.0
flake8>=5.0.0
mypy>=0.971
pylint>=2.14.0

# Monitoring and Logging
prometheus-client>=0.14.0
structlog>=22.1.0
colorlog>=6.6.0

# Production
gunicorn>=20.1.0
celery>=5.2.0
flower>=1.2.0
